changeCurrentRoom
